big refactorisation:
now, there are two columns in the result of longen:
- a character "cv" ("Character Value"), and a
- numeric column "nv" ("Nmueric Value").

cmp_all doesn't depend on cmp_cts anymore and is turned back to the methodology in a previous commit, namely:
The labels are again joint to the raw data in long format.
For non-occuring labelled values, this creates additional rows.
In these rows, 2 columns with the label values are added.
When these long dataframes are then full_joint together, this puts non-curring values in multiple dataframes in the same row.
These rows then just must not be counted when summarising.
The count n is "set" to 0 (by checking is.na() for the id column).
Now, the function cmp_no_cts_vallabs() (which was calculating exactly the rows of these non-occuring labelled values) isn't needed anymore.

The following functions have been removed:
- cmp_no_cts_vallabs
- tab_no_cts_vallabs